+++ noodles-util/src/alignment/async/io/writer.rs
@@ -79,10 +79,7 @@ where
         match self {
             Self::Sam(writer) => writer.write_alignment_record(header, record).await,
             Self::Bam(writer) => writer.write_alignment_record(header, record).await,
-            Self::Cram(writer) => {
-                let record = cram::Record::try_from_alignment_record(header, record)?;
-                writer.write_record(header, record).await
-            }
+            Self::Cram(writer) => writer.write_alignment_record(header, record).await,
         }
     }